Use text-property-(any|not-all)
This is more clear. The "property-change" functions can only jump to a
change (which could be the start or end of the desired property).
Fixes #6 and fixes #7.
Partially addresses #5 by ensuring that only the beginning of a link is
given an overlay (doesn't entirely fix because there is overlap between
htmlize links and text urls).