Get info from function called by link(s)-action
The function called by link-hint--link-action,
link-hint--multiple-links-action, and link-hint--all-links-action can
now return a list of values to determine whether the point should be
restored and to give more specific information to be messaged.

The first value of the list will prevent the position being restored if
non-nil. The second value is to be a string representing the action as
a verb such as "Copied." The third value is to be a string
representing the link (if applicable).

If the function does not return a list, the position will be restored.
If the function does not return a list or the list does not include
the second or third values, a more generic message will be printed to
the user.

Fixes #11. Fixes #10.