Fix customize widget links
Customize widgets have no category property, so `button-at' doesn't
work.

This also removes the :copy action. Maybe we could find a way to keep
it, but I couldn't really think of any use case, and in any case the
ability to invoke the widget actions is certainly more important than
treating them as text, whatever that was supposed to do.