(bug): fix minor mode check was always false
In `link-hint--var-valid-p` we're checking for active major or minor modes. For the
major-mode check we simply compared against the `major-mode` variable, but for the
minor mode check we were using the `bound-and-true-p` macro as a function. instead of
checking whether a minor mode variable was active, we were checking whether a
variable named `var` was. Which caused the condition to always be false.